# docs/content/api/openapi.yaml
openapi: 3.0.3
info:
  title: TrueBlocks API
  contact:
    email: info@trueblocks.io
    url: https://www.trueblocks.io
  license:
    name: GPL 3.0
    url: http://www.gnu.org/licenses/
  version: 0.17.0-alpha
  description: >

    A REST layer over the TrueBlocks application. With `chifra serve`, you can
    run this on your own machine, and make calls to `localhost`.

    ## How to use this API

    This API's endpoints are exact translations of the commands used by the chifra
    CLI application, and the query parameters mirror the commands' options and
    flags. If you want details, [the commands have their own documentation page](/docs/using/introducing-chifra/).

    For detailed descriptions of fields, see [the data model reference](/data-model/intro/).

      ### Before you begin

    1. [Install the trueblocks-core application](/docs/install/install-trueblocks/)
      on your machine, change your configs as needed.
    2. Run `chifra serve`


      ### Example queries

     By default, all calls are to `localhost:8080`.
     All options and flags are passed through query parameters.

     For example, to get block `100`, make a call to `/blocks` and specify
     the block you want in the query parameter:

     ```shell
     curl "http://localhost:8080/blocks?blocks=100"
     ```

     Some parameters support ranges:

     ```shell
     curl "http://localhost:8080/blocks?blocks=100-120"
     ```

...
     the unique addresses from that block range:

     ```shell
...

...

...

...

...
servers:
...
paths:
  /list:
    get:
      tags:
        - Accounts
      summary: List transactions
      description: List every appearance of an address anywhere on the chain.
      operationId: accounts-list
      parameters:
        - name: addrs
          description: one or more addresses (0x...) to list
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: count
          description: present only the number of records
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: appearances
          description: export a list of appearances
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: a single appearance of an address anywhere in the chain in the form of blockNumber.transactionIndex
                    type: array
                    items:
                      $ref: "#/components/schemas/appearance"
                example:
                  {
                    "data": [
                      {
                        "blockNumber": 8854723,
                        "transactionIndex": 61,
                        "address": "0xf503017d7baf7fbc0fff7492b751025c6a78179b",
                        "name": "TrueBlocks 2020 Wallet 0xf503"
                      },
                      {
                        "blockNumber": 8855603,
                        "transactionIndex": 121,
                        "address": "0xf503017d7baf7fbc0fff7492b751025c6a78179b",
                        "name": "TrueBlocks 2020 Wallet 0xf503"
                      },
                      {
                        "...": "..."
                      }
                    ]
                  }
        "400":
          description: bad input parameter
  /export:
    get:
      tags:
        - Accounts
      summary: Export details
      description: Export full detail of transactions for one or more addresses.
      operationId: accounts-export
      parameters:
        - name: addrs
          description: one or more addresses (0x...) to export
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: topics
          description: filter by one or more log topics (only for --logs option)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: topic
        - name: fourbytes
          description: >
             filter by one or more fourbytes (only for transactions and trace options)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: fourbyte
        - name: appearances
          description: export a list of appearances
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: receipts
          description: export receipts instead of transactional data
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: logs
          description: export logs instead of transactional data
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: traces
          description: export traces instead of transactional data
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: statements
          description: >
            export reconciliations instead of transactional data (requires --accounting option)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: neighbors
          description: export the neighbors of the given address
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: accounting
          description: >
            attach accounting records to the exported data (applies to transactions export only)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: articulate
          description: articulate transactions, traces, logs, and outputs
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: cache
          description: write transactions to the cache (see notes)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: cacheTraces
          description: write traces to the cache (see notes)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: count
          description: >
            only available for --appearances mode, if present, return only the number of records
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: firstRecord
          description: the first record to process
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: string
        - name: maxRecords
          description: the maximum number of records to process before reporting
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: string
        - name: relevant
          description: >
            for log and accounting export only, export only logs relevant to one of the given export addresses
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: emitter
          description: >
            for log export only, export only logs if emitted by one of these address(es)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: topic
          description: for log export only, export only logs with this topic(s)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: topic
        - name: clean
          description: clean (i.e. remove duplicate appearances) from all existing monitors
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: a single appearance of an address anywhere in the chain in the form of blockNumber.transactionIndex
                    type: array
                    items:
                      oneOf:
                        - $ref: "#/components/schemas/appearance"
                        - $ref: "#/components/schemas/reconciliation"
                        - $ref: "#/components/schemas/transaction"
                        - $ref: "#/components/schemas/receipt"
                        - $ref: "#/components/schemas/log"
                        - $ref: "#/components/schemas/trace"
                example:
                  {
                    "data":
                      [
                        {
                          "hash": "0x1a898c5448b37f693343917ea40b7ad1c43b28a4ddd37af1bd6d0bb4a0c99891",
                          "blockHash": "0x1d7f60b34bad8ce0ba5e49e3845c7a62ba8db440b81f8df98ec6349c4d28ca8e",
                          "blockNumber": 8854723,
                          "transactionIndex": 61,
                          "timestamp": 1572639538,
                          "from": "0xbb984f85bd52d78eb2fbf2c5598bae7abb98c5bc",
                          "to": "0xf503017d7baf7fbc0fff7492b751025c6a78179b",
                          "value": 5e+18,
                          "gas": 21000,
                          "gasPrice": 10000000000,
                          "maxFeePerGas": 0,
                          "maxPriorityFeePerGas": 0,
                          "isError": 0,
                          "hasToken": 0,
                          "receipt":
                            {
                              "contractAddress": "0x0",
                              "gasUsed": 21000,
                              "effectiveGasPrice": 10000000000,
                              "status": 1,
                            },
                          "compressedTx": "0x ( )",
                          "gasCost": 210000000000000,
                          "gasUsed": 21000,
                          "date": "2019-11-01 20:18:58 UTC",
                          "ether": 5,
                        },
                      ],
                  }
        "400":
          description: bad input parameter
  /monitors:
    get:
      tags:
        - Accounts
      summary: Manage Monitors
      description: Add, remove, clean, and list address monitors.
      operationId: accounts-monitors
      parameters:
        - name: addrs
          description: one or more addresses (0x...) to process
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: appearances
          description: export a list of appearances
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: count
          description: present only the number of records
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: clean
          description: clean (i.e. remove duplicate appearances) from monitors
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: A local file indicating a user's interest in an address. Includes caches for reconicilations, transactions and appearances as well as an optional association to named account
                    type: array
                    items:
                      $ref: "#/components/schemas/monitor"
        "400":
          description: bad input parameter
  /names:
    get:
      tags:
        - Accounts
      summary: Names
      description: Query addresses or names of well known accounts.
      operationId: accounts-names
      parameters:
        - name: terms
          description: a space separated list of one or more search terms
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: string
        - name: expand
          description: >
            expand search to include all fields (search name, address, and symbol otherwise)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: matchCase
          description: do case-sensitive search
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: all
          description: include all accounts in the search
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: custom
          description: include your custom named accounts
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: prefund
          description: include prefund accounts
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: named
          description: include well know token and airdrop addresses in the search
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: addr
          description: display only addresses in the results (useful for scripting)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: collections
          description: display collections data
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: tags
          description: export the list of tags and subtags only
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: an association between a name and address for use throughout TrueBlocks
                    type: array
                    items:
                      $ref: "#/components/schemas/name"
                example:
                  [
                    {
                      "tags": "50-Tokens:ERC20",
                      "address": "0xfe5f141bf94fe84bc28ded0ab966c16b17490657",
                      "name": "LibraToken",
                      "symbol": "LBA",
                      "source": "On chain",
                      "decimals": 18,
                      "description": "Decentralized lending infrastructure facilitating open access to credit networks on Ethereum.",
                    },
                  ]
        "400":
          description: bad input parameter
  /abis:
    get:
      tags:
        - Accounts
      summary: ABIs
      description: Fetches the ABI for a smart contract.
      operationId: accounts-abis
      parameters:
        - name: addrs
          description: a list of one or more smart contracts whose ABIs to display
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: sol
          description: extract the abi definition from the provided .sol file(s)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: find
          description: search for function or event declarations given a four- or 32-byte code(s)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: string
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: a collection of function and event definitions as described elsewhere in the Ethereum system
                    type: array
                    items:
                      $ref: "#/components/schemas/abi"
                example:
                  {
                    "data": [
                      {
                        "name": "PairCreated",
                        "type": "event",
                        "stateMutability": "nonpayable",
                        "signature": "PairCreated(address,address,address,uint256)",
                        "encoding": "0x0d3648bd0f6ba80134a33ba9275ac585d9d315f0ad8355cddefde31afa28d0e9",
                        "inputs": [
                          {
                            "type": "address",
                            "name": "token0",
                            "indexed": true,
                            "internalType": "address"
                          },
                          {
                            "type": "address",
                            "name": "token1",
                            "indexed": true,
                            "internalType": "address"
                          },
                          {
                            "type": "address",
                            "name": "pair",
                            "internalType": "address"
                          },
                          {
                            "type": "uint256",
                            "name": "val0",
                            "internalType": "uint256"
                          }
                        ]
                      }
                    ]
                  }
        "400":
          description: bad input parameter
  /blocks:
    get:
      tags:
        - Chain Data
      summary: Get blocks
      description: Retrieve one or more blocks from the chain or local cache.
      operationId: chaindata-blocks
      parameters:
        - name: blocks
          description: a space-separated list of one or more block identifiers
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: blknum
        - name: hashes
          description: >
            display only transaction hashes, default is to display full transaction detail
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: uncles
          description: display uncle blocks (if any) instead of the requested block
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: trace
          description: export the traces from the block as opposed to the block data
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: apps
          description: display a list of uniq address appearances in the block
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: uniq
          description: display a list of uniq address appearances per transaction
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: logs
          description: display only the logs found in the block(s)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: emitter
          description: >
            for the --logs option only, filter logs to show only those logs emitted by the given address(es)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: address
        - name: topic
          description: >
            for the --logs option only, filter logs to show only those with this topic(s)
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: topic
        - name: count
          description: display the number of the lists of appearances for --addrs or --uniq
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
        - name: cache
          description: force a write of the block to the cache
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: block data directly from the blockchain
                    type: array
                    items:
                      $ref: "#/components/schemas/block"
                example:
                  {
                    "blockNumber": 3141592,
                    "hash": "0x0b4c6fb75ded4b90218cf0346b0885e442878f104e1b60bf75d5b6860eeacd53",
                    "...": "...",
                    "transactions": []
                  }
        "400":
          description: bad input parameter
...
  /receipts:
    get:
      tags:
        - Chain Data
      summary: Get receipts
      description: Retrieve receipts for the given transaction(s).
      operationId: chaindata-receipts
      parameters:
        - name: transactions
          description: a space-separated list of one or more transaction identifiers
          required: true
          style: form
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
              format: tx_id
        - name: articulate
          description: articulate the retrieved data if ABIs can be found
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: receipt data for a given transaction directly from the blockchain
                    type: array
                    items:
                      $ref: "#/components/schemas/receipt"
        "400":
          description: bad input parameter
...
  /when:
    get:
      tags:
        - Chain Data
      summary: Get block dates
      description: Find block(s) based on date, blockNum, timestamp, or 'special'.
      operationId: chaindata-when
      parameters:
        - name: timestamps
          description: ignore other options and generate timestamps only
          required: false
          style: form
          in: query
          explode: true
          schema:
            type: boolean
      responses:
        "200":
          description: returns the requested data
          content:
            application/json:
              schema:
                properties:
                  data:
                    description: An indentified block named for convienience
                    type: array
                    items:
                      $ref: "#/components/schemas/datedBlock"
        "400":
          description: bad input parameter
...
components:
  schemas:
...
    abi:
      description: "a collection of function and event definitions as described elsewhere in the Ethereum system"
      type: object
      properties:
        address:
          type: string
          format: address
          description: "the smart contract that implements this abi"
        interfaces:
          type: array
          items:
            $ref: "#/components/schemas/function"
          description: "the list of events and functions on this abi"
...